SecondVer edits
Added a short paragraph on the about page with an "id" so we can style it later if we want.
Added a form to the add-recipe page with all the user inputs we would need to add the recipe.
Added some css to make the description and ingredients text boxes slightly bigger.